--- app/src/main/java/am/te/myapplication/PopulateLists.java
@@ -1,11 +1,0 @@
-package am.te.myapplication;
-
-/**
- * Created by Collin on 3/18/15.
- */
-public abstract class PopulateLists {
-
-    private static String URL;
-
-
-}
